[swift-transformers] Update secret token
I don't know what I'm doing. I thought that `HF_TOKEN` had to be defined in the source repo, but it's not being forwarded (most possibly for security reasons when people open PRs from forks). I'm here trying to reuse `HF_HUB_READ_TOKEN`, which I saw being used in other workflows. I assume it's defined as a secret of _this_ repo; if so, it should work, since the repos we need to access are all public.